add InsultListener
